Added Feeds class to get feed by id and location
