+++ src/API/Base.php
@@ -481,90 +481,6 @@ class Base {
 
 
 	/**
-	 * Get a specific merchant feed using the given arguments.
-	 *
-	 * @param string $merchant_id The merchant ID the feed belongs to.
-	 * @param string $feed_id     The ID of the feed.
-	 *
-	 * @return mixed
-	 *
-	 * @throws \Exception PHP Exception.
-	 */
-	public static function get_merchant_feed( $merchant_id, $feed_id ) {
-		try {
-
-			$feeds = self::get_merchant_feeds( $merchant_id );
-
-			if ( 'success' !== $feeds['status'] ) {
-				throw new \Exception( esc_html__( 'Could not get feed info.', 'pinterest-for-woocommerce' ) );
-			}
-
-			if ( ! is_array( $feeds['data'] ) ) {
-				throw new \Exception( esc_html__( 'Wrong feed info.', 'pinterest-for-woocommerce' ) );
-			}
-
-			foreach ( $feeds['data'] as $feed_profile ) {
-
-				if ( $feed_id === $feed_profile->id ) {
-					return $feed_profile;
-				}
-			}
-
-			// No feed found.
-			throw new \Exception( esc_html__( 'No feed found with the requested ID.', 'pinterest-for-woocommerce' ) );
-
-		} catch ( \Exception $e ) {
-
-			Logger::log( $e->getMessage(), 'error' );
-
-			throw $e;
-		}
-	}
-
-
-	/**
-	 * Get merchant's feed based on feed location
-	 *
-	 * @param string $merchant_id   The merchant ID.
-	 * @param string $feed_location The feed full location.
-	 *
-	 * @return mixed
-	 *
-	 * @throws \Exception PHP Exception.
-	 */
-	public static function get_merchant_feed_by_location( $merchant_id, $feed_location ) {
-		try {
-
-			$feeds = self::get_merchant_feeds( $merchant_id );
-
-			if ( 'success' !== $feeds['status'] ) {
-				throw new \Exception( esc_html__( 'Could not get feed info.', 'pinterest-for-woocommerce' ) );
-			}
-
-			if ( ! is_array( $feeds['data'] ) ) {
-				throw new \Exception( esc_html__( 'Wrong feed info.', 'pinterest-for-woocommerce' ) );
-			}
-
-			foreach ( $feeds['data'] as $feed_profile ) {
-
-				if ( $feed_location === $feed_profile->location_config->full_feed_fetch_location ) {
-					return $feed_profile;
-				}
-			}
-
-			// No feed found.
-			throw new \Exception( esc_html__( 'No feed found with the requested location.', 'pinterest-for-woocommerce' ) );
-
-		} catch ( \Exception $e ) {
-
-			Logger::log( $e->getMessage(), 'error' );
-
-			throw $e;
-		}
-	}
-
-
-	/**
 	 * Get a merchant's feeds.
 	 *
 	 * @param string $merchant_id The merchant ID the feed belongs to.
